move logic to utilities
